TradingBot_Pro_2025/requirements.txt
# Web Framework
flask==3.1.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.26.0,<2.0.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.15.0
joblib>=1.3.0

# Financial Data & Analysis
yfinance>=0.2.0
ccxt>=4.0.0
ta>=0.10.0
python-binance>=1.0.0
websocket-client==1.6.3

# Async & Networking
aiohttp>=3.10.11

# Utilities
requests==2.32.5
python-dotenv==1.1.1
celery==5.3.4
redis==5.2.1
schedule==1.2.0
psutil==5.9.6
colorama==0.4.6
click==8.1.7
tqdm==4.66.1

# Database
sqlalchemy==2.0.36

# Visualization & UI
matplotlib>=3.6.0
plotly>=5.0.0
seaborn>=0.12.0

# Advanced ML
lightgbm>=4.0.0
xgboost>=1.7.0

# Streamlit & Dashboard
streamlit>=1.20.0
dash>=2.10.0

# Notifications
python-telegram-bot==20.6
discord.py==2.3.2

# Testing
pytest==7.4.4
pytest-mock==3.14.0
pytest-flask==1.3.0
pytest-asyncio==0.21.1
coverage==7.6.9

# Production
gunicorn==22.0.0
